+++ packages/registration_delivery/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.0.3+3
+* Updated Service model additionalDetails objectType in digit_data_model package
+
+## 1.0.3+2
+
+* Bug Fix -
+  * fixed wrong status map of beneficiaryRefused in utils
+* Sorting tasks based on createTime
+
+## 1.0.4-dev.1
+
+* Integrated survey_form package
+
 ## 1.0.3+1
 
 * Bug Fixes -